Change BCT branch, fix changing branches
This changes the way the install script updates the Git repository so
that it first fetches updates from the server, then switches to the
configured branch, and then merges in any changes from origin.

This change should allow branch changes to work both from deleted
branches and to newly created branches in the future.